+++ CHANGELOG.md
@@ -2,6 +2,89 @@
 
 <!-- version list -->
 
+## v0.13.1 (2025-08-05)
+
+### Bug Fixes
+
+- Correct import statement for Bucket class
+  ([`cbd5f6d`](https://github.com/TimKleindick/general_manager/commit/cbd5f6d990dc59dc05d5229bb3714621ffb188d8))
+
+- Ensure path existence check in PathMap methods
+  ([`ad5700c`](https://github.com/TimKleindick/general_manager/commit/ad5700ceb03ff95951061eb77585476c718ce6fb))
+
+- Handle None attribute types in PathTracer connections
+  ([`bff43f2`](https://github.com/TimKleindick/general_manager/commit/bff43f267e44a16acd16c1a0d82c8b375fb50918))
+
+- Simplify path tracer retrieval logic in PathMap
+  ([`bd6e327`](https://github.com/TimKleindick/general_manager/commit/bd6e327fb0253697b38a9de353d81b8d9061a0fb))
+
+### Refactoring
+
+- Clean up whitespace and improve path mapping documentation
+  ([`e0784cf`](https://github.com/TimKleindick/general_manager/commit/e0784cf43231381c11a7e7d184e2a75d6e9be814))
+
+- Clean up whitespace and improve readability in PathMapping unit tests
+  ([`e796941`](https://github.com/TimKleindick/general_manager/commit/e7969414007459eaa956e879df0ee369d2e58263))
+
+- Reorganize imports and simplify test setup in pathMapping tests
+  ([`c5656e6`](https://github.com/TimKleindick/general_manager/commit/c5656e685e7d990b74ec3d1d44004764ef38fa1b))
+
+### Testing
+
+- Add integration and unit tests for PathMap functionality
+  ([`fb0e080`](https://github.com/TimKleindick/general_manager/commit/fb0e080036ff787e087c38b6448a99a5506dc810))
+
+
+## v0.13.0 (2025-08-03)
+
+### Bug Fixes
+
+- Apply permission filters to queryset in GraphQL class
+  ([`5125e33`](https://github.com/TimKleindick/general_manager/commit/5125e33f494fbee969d90f3af172e04b9d264d86))
+
+- Improve type checking for based_on object in ManagerBasedPermission
+  ([`95c483e`](https://github.com/TimKleindick/general_manager/commit/95c483e70a41e591b945505f5d1f02e30a57e014))
+
+- Refactor permission attributes initialization and handle based_on condition
+  ([`2e21b3b`](https://github.com/TimKleindick/general_manager/commit/2e21b3b64619942de5a011868f2600107f7245c3))
+
+- Remove unnecessary blank lines in test_calculationBucket.py
+  ([`3117b2a`](https://github.com/TimKleindick/general_manager/commit/3117b2afcf7b789a7a8cdd1d3e24f691710bdddc))
+
+- Remove unused import statement and clean up
+  ([`784c174`](https://github.com/TimKleindick/general_manager/commit/784c1742dbe304bd60f7c881c911e2641ca33c0d))
+
+- Remove unused import statement in GraphQL class
+  ([`8bab0a3`](https://github.com/TimKleindick/general_manager/commit/8bab0a354fec0c94b6f21ebfd3d1a13b5ac386cf))
+
+- Remove unused permission inheritance test method
+  ([`b987be3`](https://github.com/TimKleindick/general_manager/commit/b987be3cedfd58381acb48ca8575cb77f548acef))
+
+- Update condition to check for None or NOT_PROVIDED in many-to-many kwargs
+  ([`ca779b6`](https://github.com/TimKleindick/general_manager/commit/ca779b616f208cf28ddf1a258483f151337fbdce))
+
+### Features
+
+- Add 'matches' permission function to permission_functions
+  ([`542e820`](https://github.com/TimKleindick/general_manager/commit/542e82056a5b5085ae75359d63d6066c82647fd0))
+
+- Add `none` method to Bucket class and import QuerySet
+  ([`ca3c8b5`](https://github.com/TimKleindick/general_manager/commit/ca3c8b5b5b14099f8789fc6914216e81def42738))
+
+- Refactor permission initialization and add difference between None and no result
+  ([`94d4809`](https://github.com/TimKleindick/general_manager/commit/94d480934097bd51106ad9d04e702e7887466e3c))
+
+### Refactoring
+
+- Clean up code formatting and improve permission validation tests
+  ([`2070a60`](https://github.com/TimKleindick/general_manager/commit/2070a608306cbddd92a53e0cceec3a09808f93bc))
+
+### Testing
+
+- Add integration tests for manager-based permissions and family updates
+  ([`54eebde`](https://github.com/TimKleindick/general_manager/commit/54eebde4906d1386a731f9c336686803aeb535de))
+
+
 ## v0.12.2 (2025-07-28)
 
 ### Bug Fixes